[Refactor] 풀더구조 domain / global 로 변경
